+++ api/python/src/PDB/init.cpp
@@ -5,6 +5,7 @@
 #include <nanobind/stl/string.h>
 
 #include "LIEF/PDB/DebugInfo.hpp"
+#include "LIEF/PDB/BuildMetadata.hpp"
 #include "LIEF/PDB/Type.hpp"
 
 namespace LIEF::pdb::py {
@@ -17,6 +18,7 @@ void init(nb::module_& m) {
     )doc"_doc, "path"_a
   );
 
+  create<LIEF::pdb::BuildMetadata>(pdb);
   create<LIEF::pdb::Type>(pdb);
   create<LIEF::pdb::DebugInfo>(pdb);
   create<LIEF::pdb::PublicSymbol>(pdb);