+++ api/rust/cargo/lief/src/pdb/function.rs
@@ -47,6 +47,12 @@ impl Function<'_> {
     }
 }
 
+impl std::fmt::Display for Function<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.ptr.to_string())
+    }
+}
+
 declare_fwd_iterator!(
     Functions,
     Function<'a>,