+++ include/LIEF/PDB/CompilationUnit.hpp
@@ -25,6 +25,7 @@
 
 namespace LIEF {
 namespace pdb {
+class BuildMetadata;
 
 namespace details {
 class CompilationUnit;
@@ -116,6 +117,19 @@ class LIEF_API CompilationUnit {
   /// If the PDB does not contain or has an empty DBI stream, it returns
   /// an empty iterator.
   function_iterator functions() const;
+
+  /// Return build metadata such as the version of the compiler or
+  /// the original source language of this compilation unit
+  std::unique_ptr<BuildMetadata> build_metadata() const;
+
+  std::string to_string() const;
+
+  LIEF_API friend
+    std::ostream& operator<<(std::ostream& os, const CompilationUnit& CU)
+  {
+    os << CU.to_string();
+    return os;
+  }
 
   private:
   std::unique_ptr<details::CompilationUnit> impl_;