+++ include/LIEF/PDB/PublicSymbol.hpp
@@ -108,6 +108,15 @@ class LIEF_API PublicSymbol {
   /// This function returns 0 if the RVA can't be computed.
   uint32_t RVA() const;
 
+  std::string to_string() const;
+
+  LIEF_API friend
+    std::ostream& operator<<(std::ostream& os, const PublicSymbol& sym)
+  {
+    os << sym.to_string();
+    return os;
+  }
+
   private:
   std::unique_ptr<details::PublicSymbol> impl_;
 };